[benchmark] Refactor profiling onto a shared case runtime
- Measurement loop and DevTools profiling both drive one createCaseRuntime
  (per-case mount/interact?/unmount/isMounted state machine); extract the
  interactive panel into profileSession.tsx (pure DOM, injected runtime).
- Move BenchmarkCaseRuntime to types.ts; runtime plants the default Element
  Timing sentinel; measurement pre-wraps renderFn in BenchProfiler, profile
  renders bare.
- Collapse the two launch-arg sets into one minimal LAUNCH_ARGS; drop the V8
  determinism flags / --disable-gpu (a sweep showed they don't reduce variance
  and --no-opt costs ~40% render speed).
- profile?: boolean + viewport?: { width, height } with a single 1920x1080
  default for both modes; env BENCHMARK_PROFILE / BENCHMARK_VIEWPORT.
